+++ bciers/apps/compliance/src/app/components/compliance-summary/request-issuance/internal/review-compliance-summary/ComplianceSummaryReviewPage.tsx
@@ -1,11 +1,10 @@
 import {
-  generateRequestIssuanceTaskList,
+  generateIssuanceRequestTaskList,
   ActivePage,
-} from "@/compliance/src/app/components/taskLists/requestIssuanceTaskList";
+} from "@/compliance/src/app/components/taskLists/internal/issuanceRequestTaskList";
 import CompliancePageLayout from "@/compliance/src/app/components/layout/CompliancePageLayout";
-import ComplianceSummaryReviewComponent from "@/compliance/src/app/components/compliance-summary/request-issuance/review-compliance-summary/ComplianceSummaryReviewComponent";
+import ComplianceSummaryReviewComponent from "./ComplianceSummaryReviewComponent";
 import { getRequestIssuanceComplianceSummaryData } from "@/compliance/src/app/utils/getRequestIssuanceComplianceSummaryData";
-import { RequestIssuanceComplianceSummaryData } from "@/compliance/src/app/types";
 import { IssuanceStatus } from "@bciers/utils/src/enums";
 import { redirect } from "next/navigation";
 
@@ -16,23 +15,21 @@ interface Props {
 export default async function ComplianceSummaryReviewPage({
   compliance_summary_id: complianceSummaryId,
 }: Readonly<Props>) {
-  const complianceSummary: RequestIssuanceComplianceSummaryData =
+  const complianceSummary =
     await getRequestIssuanceComplianceSummaryData(complianceSummaryId);
 
   // Redirect the user to track status page if user already requested issuance or issuance is approved or declined
   if (
-    [
-      IssuanceStatus.ISSUANCE_REQUESTED,
-      IssuanceStatus.APPROVED,
-      IssuanceStatus.DECLINED,
-    ].includes(complianceSummary.issuance_status as IssuanceStatus)
+    [IssuanceStatus.APPROVED, IssuanceStatus.DECLINED].includes(
+      complianceSummary.issuance_status,
+    )
   ) {
     redirect(
       `/compliance-summaries/${complianceSummaryId}/track-status-of-issuance`,
     );
   }
 
-  const taskListElements = generateRequestIssuanceTaskList(
+  const taskListElements = generateIssuanceRequestTaskList(
     complianceSummaryId,
     complianceSummary.reporting_year,
     ActivePage.ReviewComplianceSummary,
