+++ bciers/apps/compliance/src/app/idir/cas_admin/compliance-summaries/[compliance_summary_id]/(request-issuance)/request-issuance-review-summary/page.tsx
@@ -1,4 +1,4 @@
 import defaultPageFactory from "@bciers/components/nextPageFactory/defaultPageFactory";
-import Page from "@/compliance/src/app/components/compliance-summary/request-issuance/review-compliance-summary/ComplianceSummaryReviewPage";
+import Page from "@/compliance/src/app/components/compliance-summary/request-issuance/internal/review-compliance-summary/ComplianceSummaryReviewPage";
 
 export default defaultPageFactory(Page);
